requirements.txt
Flask==3.0.0
flask-login==0.6.3
Werkzeug==3.0.1
psycopg2-binary
python-dotenv==1.0.0
gunicorn==21.2.0
